Moved Wayland code to the correct folder, fill gaps found so health can be read more often
